resin-ntp-config: Resolve NTP source to IP address before addition
When using NTP pools we might be passing different NTP sources to the add
and the burst commands as the pool will resolve to a different IP address.

This commit will first resolve the NTP source to an IP address and then
use it to add the server and perform the burst.

Fixes #1903

Changelog-entry: Resolve NTP sources to IP addressed before using them
Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>